+++ SportQuestTracker/SportQuestTracker.csproj
@@ -35,4 +35,8 @@
     </Content>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Data" />
+  </ItemGroup>
+
 </Project>
