+++ SportQuestTracker/appsettings.json
@@ -7,8 +7,8 @@
     }
   },
   "ConnectionStrings": {
-    "DefaultConnection": "Server=.\\SQLEXPRESS\\mssqllocaldb;Database=SportQuestTracker;Trusted_Connection=True;MultipleActiveResultSets=true",
-    "SportQuestTrackerContext": "Server=(localdb)\\mssqllocaldb;Database=SportQuestTrackerContext-c290ce74-595a-49ee-8125-8d6675e9a720;Trusted_Connection=True;MultipleActiveResultSets=true"
+    "DefaultConnection":"server=localhost; user id=pat;password=345Baksio*;database=SportQuestTracker",
+    "SportQuestTrackerContext": "Server=(localdb)\\localhost;Database=SportQuestTrackerContext-c290ce74-595a-49ee-8125-8d6675e9a720;Trusted_Connection=True;MultipleActiveResultSets=true"
   },
   "AllowedHosts": "*"
-}
+  }
